chore: restructuring and docs
