+++ apps/docs/app/(docs)/api-reference/page.mdx
@@ -1,24 +1,6 @@
 # API Reference
 
-This is the reference for the utility functions that `react-tweet` provides for [building your own tweet components](/custom-theme) or simply fetching a tweet. Navigate to the docs for the [Twitter theme](/twitter-theme) if you want to render the existing Tweet components instead.
-
-## `getTweet`
-
-```tsx
-import { getTweet, type Tweet } from "react-tweet/api";
-
-function getTweet(
-  id: string,
-  fetchOptions?: RequestInit
-): Promise<Tweet | undefined>;
-```
-
-Fetches and returns a [`Tweet`](https://github.com/vercel-labs/react-tweet/blob/main/packages/react-tweet/src/api/types/tweet.ts). It accepts the following params:
-
-- **id** - `string`: the tweet ID. For example in `https://twitter.com/chibicode/status/1629307668568633344` the tweet ID is `1629307668568633344`.
-- **fetchOptions** - `RequestInit` (Optional): options to pass to [`fetch`](https://developer.mozilla.org/en-US/docs/Web/API/fetch).
-
-If a tweet is not found it returns `undefined`.
+This is the reference for the utility functions that `bsky-react-post` provides for [building your own post components](/custom-theme) or simply fetching a post. Navigate to the docs for the [Bluesky theme](/bluesky-theme) if you want to render the existing Post components instead.
 
 ## `fetchTweet`
 
@@ -38,18 +20,6 @@ Fetches and returns a [`Tweet`](https://github.com/vercel-labs/react-tweet/blob/
 - **data** - `Tweet` (Optional): The tweet data.
 - **tombstone** - `true` (Optional): Indicates if the tweet has been made private.
 - **notFound** - `true` (Optional): Indicates if the tweet was not found.
-
-## `enrichTweet`
-
-```tsx
-import { enrichTweet, type EnrichedTweet } from "react-tweet";
-
-const enrichTweet: (tweet: Tweet) => EnrichedTweet;
-```
-
-Enriches a [`Tweet`](https://github.com/vercel-labs/react-tweet/blob/main/packages/react-tweet/src/api/types/tweet.ts) as returned by [`getTweet`](#gettweet) with additional data. This is useful to more easily build custom tweet components.
-
-It returns an [`EnrichedTweet`](https://github.com/vercel-labs/react-tweet/blob/main/packages/react-tweet/src/utils.ts).
 
 ## `useTweet`
