+++ apps/docs/app/(docs)/contributing/page.mdx
@@ -1,7 +1,7 @@
 # Contributing
 
-To contribute, clone the [`react-tweet` repository](https://github.com/vercel-labs/react-tweet) and run the [Next.js test app](/next#running-the-test-app) to start an app locally that uses the [`react-tweet` package](https://github.com/vercel-labs/react-tweet/blob/main/packages/react-tweet). Any changes you make to the package will be reflected in the test app.
+To contribute, clone the [`bsky-react-post` repository](https://github.com/rhinobase/react-bluesky) and run the [Next.js test app](/next#running-the-test-app) to start an app locally that uses the [`bsky-react-post` package](https://github.com/rhinobase/react-bluesky/tree/main/packages/core). Any changes you make to the package will be reflected in the test app.
 
-Once you're done making changes, [submit a pull request](https://github.com/vercel-labs/react-tweet/compare).
+Once you're done making changes, [submit a pull request](https://github.com/rhinobase/react-bluesky/compare).
 
-It's recommended to [open an issue](https://github.com/vercel-labs/react-tweet/issues/new) first before making any major changes to the package so that we can discuss the changes before you start working on them.
+It's recommended to [open an issue](https://github.com/rhinobase/react-bluesky/issues/new) first before making any major changes to the package so that we can discuss the changes before you start working on them.
