--- apps/docs/app/(docs)/create-react-app/page.mdx
@@ -1,31 +1,0 @@
-# Create React App
-
-## Installation
-
-Follow the [installation docs in the Introduction](/#installation).
-
-## Usage
-
-In any component, import `Tweet` from `react-tweet` and use it like so:
-
-```tsx
-import { Tweet } from "react-tweet";
-
-export default function App() {
-  return <Tweet id="1628832338187636740" />;
-}
-```
-
-You can learn more about `Tweet` in the [Twitter theme docs](/twitter-theme).
-
-## Running the test app
-
-Clone the [`react-tweet`](https://github.com/vercel-labs/react-tweet) repository and then run the following command:
-
-```bash
-pnpm install && pnpm dev --filter=create-react-app...
-```
-
-The app will be up and running at (localhost:3002)[http://localhost:3002] for the [CRA example](https://github.com/vercel-labs/react-tweet/tree/main/apps/create-react-app).
-
-The source code for `react-tweet` is imported from [packages/react-tweet](https://github.com/vercel-labs/react-tweet/tree/main/packages/react-tweet) and any changes you make to it will be reflected in the app immediately.
