+++ package.json
@@ -48,7 +48,7 @@
     "@next/mdx": "^15.0.3",
     "@rafty/ui": "^1.7.7",
     "@sindresorhus/slugify": "^2.2.1",
-    "@tailwindcss/typography": "^0.5.9",
+    "@tailwindcss/typography": "0.5.9",
     "@types/mdx": "^2.0.13",
     "acorn": "^8.14.0",
     "edge-cors": "^0.2.1",
