--- packages/core/src/components/PostContainer/post-container.module.css
@@ -1,25 +1,0 @@
-.root {
-  width: 100%;
-  min-width: 250px;
-  max-width: 550px;
-  /* Base font styles */
-  color: var(--post-font-color);
-  font-family: var(--post-font-family);
-  font-weight: 400;
-  box-sizing: border-box;
-  border: var(--post-border);
-  border-radius: 0.75rem;
-  margin: var(--post-container-margin);
-  background-color: var(--post-bg-color);
-  transition-property: background-color, box-shadow;
-  transition-duration: 0.2s;
-}
-
-.root:hover {
-  background-color: var(--post-bg-color-hover);
-}
-.article {
-  position: relative;
-  box-sizing: inherit;
-  padding: 0.75rem 1rem;
-}
